Move authenticate method into new db.js file.
